增加优惠券数据库迁移文件

+++ database/migrations/2015_11_18_132512_add_polymorphic_columns_to_uploads.php
@@ -12,15 +12,17 @@ class AddPolymorphicColumnsToUploads extends Migration
      */
     public function up()
     {
-        Schema::table('uploads', function (Blueprint $table) {
-            $table->integer('uploadable_id')->nullable();
-            $table->string('uploadable_type')->nullable();
+        if (!Schema::hasTable('uploads')) {
+            Schema::table('uploads', function (Blueprint $table) {
+                $table->integer('uploadable_id')->nullable();
+                $table->string('uploadable_type')->nullable();
 
-            $table->index([
-                'uploadable_id',
-                'uploadable_type'
-            ],'uploadable_index');
-        });
+                $table->index([
+                    'uploadable_id',
+                    'uploadable_type'
+                ],'uploadable_index');
+            });
+        }
     }
 
     /**
@@ -30,12 +32,14 @@ class AddPolymorphicColumnsToUploads extends Migration
      */
     public function down()
     {
-        Schema::table('uploads', function (Blueprint $table) {
-            $table->dropIndex('uploadable_index');
-        });
-        Schema::table('uploads', function (Blueprint $table) {
-            $table->dropColumn('uploadable_id');
-            $table->dropColumn('uploadable_type');
-        });
+        if (!Schema::hasTable('uploads')) {
+            Schema::table('uploads', function (Blueprint $table) {
+                $table->dropIndex('uploadable_index');
+            });
+            Schema::table('uploads', function (Blueprint $table) {
+                $table->dropColumn('uploadable_id');
+                $table->dropColumn('uploadable_type');
+            });
+        }
     }
 }
